Renamed modules, project and refactored the code, before going on implementing CLI in the Interface module. At this point you might need to reimport the solution into your IDE
