+++ BusinessLogic/src/main/java/dk/easv/SearchEngine/BusinessLogicLayer/IndexTaskCallable.java
@@ -1,0 +1,103 @@
+package dk.easv.SearchEngine.BusinessLogicLayer;
+
+import dk.easv.SearchEngine.DataAccessLayer.Database.ContainsRepository;
+import dk.easv.SearchEngine.DataAccessLayer.Database.DocumentsRepository;
+import dk.easv.SearchEngine.DataAccessLayer.Database.TermsRepository;
+import dk.easv.SearchEngine.DataAccessLayer.FileSystem.FileLoader;
+import dk.easv.SearchEngine.DomainModels.Document;
+import dk.easv.SearchEngine.DomainModels.Term;
+
+import java.nio.file.Path;
+import java.sql.SQLException;
+import java.util.Date;
+import java.util.List;
+import java.util.concurrent.Callable;
+import java.util.stream.Collectors;
+
+public class IndexTaskCallable implements Callable {
+
+    private static final int ERROR_CODE = -1;
+
+    private final SynchronizedTermsMap syncTermsMap;
+    private final Path filePath;
+    private final FileLoader fileLoader;
+    private final TermSplitter termSplitter;
+    private final DocumentsRepository documentsRepository;
+    private final TermsRepository termsRepository;
+    private final ContainsRepository containsRepository;
+
+    public IndexTaskCallable(
+            Path filePath,
+            SynchronizedTermsMap synchronizedTermsMap,
+            FileLoader fileLoader,
+            TermSplitter termSplitter,
+            DocumentsRepository documentsRepository,
+            TermsRepository termsRepository,
+            ContainsRepository containsRepository) {
+
+        this.filePath = filePath;
+        this.syncTermsMap = synchronizedTermsMap;
+        this.fileLoader = fileLoader;
+        this.termSplitter = termSplitter;
+        this.documentsRepository = documentsRepository;
+        this.termsRepository = termsRepository;
+        this.containsRepository = containsRepository;
+    }
+
+    @Override
+    public Object call() throws Exception {
+        long startTime = System.nanoTime();
+        List<String> lines = fileLoader.loadLines(filePath);
+        Date indexTime = new Date();
+        List<Term> terms = termSplitter.splitLines(lines);
+
+        int documentId = saveDocument(filePath.toAbsolutePath().toString(), indexTime);
+        if (documentId != ERROR_CODE) {
+            List<Integer> termIDs = saveTerms(terms);
+            saveRelation(documentId, termIDs);
+            double executionTimeInSeconds = (System.nanoTime() - startTime) / 1E9;
+            return String.format("Success - Document inserted in the database! FilePath: %s.\n" +
+                            "TermIds Size: %d - Terms Size: %d.\n" +
+                            "Indexing time:%s Seconds  (Loading & Splitting lines/Saving Documents,Terms and Relations)\n"
+                    , filePath.toString(), termIDs.size(), terms.size(), executionTimeInSeconds);
+        }
+        return String.format("Error - Document was not inserted in the database! File Path: %s.\n", filePath.toString());
+    }
+
+    private int saveDocument(String absoluteFilePath, Date indexTime) {
+        Document document = new Document(absoluteFilePath, indexTime);
+        try {
+            return documentsRepository.saveDocument(document);
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return ERROR_CODE;
+        }
+    }
+
+    public List<Integer> saveTerms(List<Term> terms) {
+        for (Term term : terms) {
+            syncTermsMap.lock.lock();
+            try {
+                int mapTermId = syncTermsMap.getTermID(term.getTerm_Value());
+                if (mapTermId != syncTermsMap.TERM_NOT_IN_MAP_ERROR) {
+                    term.setTerm_ID(mapTermId);
+                } else try {
+                    int termId = termsRepository.saveTerm(term.getTerm_Value());
+                    if (termId != ERROR_CODE) {
+                        term.setTerm_ID(termId);
+                        syncTermsMap.putTerm(term);
+                    }
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+            } finally {
+                syncTermsMap.lock.unlock();
+            }
+        }
+        return terms.stream().map(Term::getTerm_ID).collect(Collectors.toList());
+    }
+
+    private void saveRelation(int documentId, List<Integer> termIDs) {
+        containsRepository.bulkSaveIndexInContainTbl(termIDs, documentId);
+    }
+}
