+++ BusinessLogic/src/main/java/dk/easv/SearchEngine/BusinessLogicLayer/TermSplitter.java
@@ -1,0 +1,11 @@
+package dk.easv.SearchEngine.BusinessLogicLayer;
+
+import dk.easv.SearchEngine.DomainModels.Term;
+
+import java.util.List;
+
+public interface TermSplitter {
+
+    List<Term> splitLines(List<String> lines);
+
+}
