+++ DataAccessLogic/src/main/java/dk/easv/SearchEngine/DataAccessLayer/Database/SearchDB.java
@@ -1,0 +1,37 @@
+package dk.easv.SearchEngine.DataAccessLayer.Database;
+
+public class SearchDB {
+
+//    public Multimap<String, String> getSimilarTermsWithDocumentPath(String searchTerm) {
+//        Multimap<String, String> termMultimap = ArrayListMultimap.create();
+//        String sqlSelect
+//                = "SELECT terms_value, documents_url\n"
+//                + "FROM terms_tbl\n"
+//                + "inner join contain_tbl\n"
+//                + "on contain_tbl.terms_id = terms_tbl.terms_id\n"
+//                + "inner join documents_tbl\n"
+//                + "on contain_tbl.documents_id = documents_tbl.documents_id\n"
+//                + "\n"
+//                + "where terms_value like ?;";
+//
+//        try (Connection connection = Database.getConnection();
+//             PreparedStatement preparedStatement
+//                = connection.prepareStatement(sqlSelect);
+//             ResultSet resultSet = preparedStatement.executeQuery();) {
+//
+//            preparedStatement.setString(1, "%" + searchTerm + "%");
+//
+//            while (resultSet.next()) {
+//                String term = resultSet.getString("terms_value");
+//                String doc_Path = resultSet.getString("documents_url");
+//                termMultimap.put(term, doc_Path);
+//            }
+//            return termMultimap;
+//
+//        } catch (SQLException ex) {
+//            Logger.getLogger(SearchDB.class.getName())
+//                    .log(Level.SEVERE, null, ex);
+//            return null;
+//        }
+//    }
+}
