--- Interface/src/main/java/com/enron/SearchEngine/FXApp.java
@@ -1,29 +1,0 @@
-package com.enron.SearchEngine;
-
-import java.net.URL;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-public class FXApp extends Application {
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        URL resource = getClass().getResource("/FXML/SearchView.fxml");
-        Parent root = FXMLLoader.load(resource);
-        Scene scene = new Scene(root);
-
-        stage.setScene(scene);
-        stage.show();
-    }
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-}
